tools/risk-assessment-generator/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "risk-assessment-generator"
version = "0.1.0"
description = "A risk assessment generator tool"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
dependencies = ["anthropic>=0.18.0", "flask>=3.0.0", "python-docx>=1.1.0"]

[project.optional-dependencies]
dev = ["pytest>=7.0.0"]

[project.scripts]
risk-assess = "risk_assessment_generator.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
